+++ audiomate/utils/download.py
@@ -3,6 +3,7 @@ This module contains any functionality for downloading and extracting data from 
 """
 
 import zipfile
+import tarfile
 import requests
 
 
@@ -21,7 +22,15 @@ def download_file(url, target_path):
 
 def extract_zip(zip_path, target_folder):
     """
-    Extract the content of the zip-file at `zip_path` into `target_folder``
+    Extract the content of the zip-file at `zip_path` into `target_folder`.
     """
     with zipfile.ZipFile(zip_path) as archive:
+        archive.extractall(target_folder)
+
+
+def extract_tar(tar_path, target_folder):
+    """
+    Extract the content of the tar-file at `tar_path` into `target_folder`.
+    """
+    with tarfile.open(tar_path, 'r') as archive:
         archive.extractall(target_folder)